Refrection and reflection
For transparent object calculations have been added for reflection and refraction.
The calculations make use of 2 recursive rays for which the recursion level is increased by 1.
The fromulas used are snell's law and fersnel equasion.
To better see the recursive rays when calling drawray now also draws a darker ray if it is behind an object.